+++ app/src/main/java/com/axfex/dorkout/data/source/local/ExercisesDao.java
@@ -22,13 +22,13 @@ import static android.arch.persistence.room.OnConflictStrategy.REPLACE;
 public interface ExercisesDao {
 
     @Query("SELECT * FROM Exercise WHERE id = :id")
-    LiveData<Exercise> getExercise(int id);
+    LiveData<Exercise> getExercise(Long id);
 
     @Query("SELECT * from Exercise where workoutId = :workoutId")
-    LiveData<List<Exercise>> getExercises(final int workoutId);
+    LiveData<List<Exercise>> getExercises(final Long workoutId);
 
     @Query("SELECT COUNT(*) from Exercise  where workoutId = :workoutId")
-    LiveData<Integer> getExercisesCount(final int workoutId);
+    LiveData<Integer> getExercisesCount(final Long workoutId);
 
     //Test stuff
     @Query("SELECT * from Exercise")
