+++ app/src/main/java/com/axfex/dorkout/views/exercises/addedit/AddEditExerciseActivity.java
@@ -21,7 +21,7 @@ public class AddEditExerciseActivity extends BaseActivity {
         Intent i = getIntent();
         if (i.hasExtra(WORKOUT_ID)) {
 
-            int workoutId = i.getIntExtra(WORKOUT_ID, 0);
+            Long workoutId = i.getLongExtra(WORKOUT_ID, 0L);
 
             FragmentManager fragmentManager = getSupportFragmentManager();
             AddEditExerciseFragment addEditExerciseFragment = (AddEditExerciseFragment) fragmentManager.findFragmentByTag(ADD_EDIT_FRAG);
